requirements.txt
numpy>=1.21.0
pandas>=1.3.0
scikit-learn>=0.24.2
opencv-python>=4.5.3
tensorflow>=2.8.0
fastapi>=0.68.0
uvicorn>=0.15.0
python-multipart>=0.0.5
websockets>=10.0
librosa>=0.8.1
pydantic>=1.8.2
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
redis>=4.0.2
pytest>=6.2.5
httpx>=0.23.0
python-dotenv>=0.19.0
tflite-runtime>=2.8.0
